+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # main
 
+# 0.4.0
+
 - Add M1 builds to release.
 - Support installing from Nix flake.
 - Add `-E/--exclude PATTERN` option to exclude paths from results.